// cardano/Types.ts
export type FaucetMetadata = {
    datumInt: string;
    policyId: string;
    tokenName: string;
    contractAddress: string;
    withdrawalAmount: string;
}

export type FaucetMetadataResult = {
    datumInt: string;
    policyId: string;
    tokenName: string;
    contractAddress: string;
    withdrawalAmount: string;
    registrationHash: string;
}